RAG-Live-Session/requirements.txt
langchain==1.2.0
python-dotenv==1.2.1
langchain-google-genai==4.1.2
sqlmodel>=0.0.14
psycopg2-binary==2.9.11
langgraph-checkpoint-postgres==3.0.2
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0
bs4==0.0.2
pysqlite3==0.6.0
pypdf
unstructured